pyproject.toml
[project]
name = "auto-posting"
version = "0.1.0"
description = "Telegram Bot API 기반 FastMCP 서버 — 메시지 전송·조회 및 채팅 관리 도구 모음"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
    "telethon>=1.39.0",
]
